Moved testFixtures to a separate module
