--- plugin/src/testFixtures/java/base/PerlTestEnvironmentTestCase.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright 2015-2024 Alexandr Evstigneev
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package base
-
-import com.intellij.ide.plugins.PluginManager
-import org.jetbrains.annotations.NonNls
-import org.junit.Test
-
-abstract class PerlTestEnvironmentTestCase : PerlLightTestCaseBase() {
-  override fun getFileExtension(): String? = "none"
-
-  override fun computeAnswerFileName(appendix: String): String = computeAnswerFileNameWithoutExtension("") + ".txt"
-
-  override fun getBaseDataPath(): @NonNls String? = "plugins"
-
-  @Test
-  fun testPluginsList() =
-    assertSameLinesWithFile(
-      testResultsFilePath,
-      PluginManager.getLoadedPlugins().map { it.pluginId.idString }.sorted().joinToString("\n")
-    )
-}
