+++ plugin/testFixtures/src/testFixtures/java/org/jetbrains/plugins/terminal/fixture/TestShellSession.java
