+++ settings.gradle.kts
@@ -63,6 +63,8 @@ include(":plugin:berrybrew")
 project(":plugin:berrybrew").projectDir = file("plugin/berrybrew")
 include(":plugin:terminal")
 project(":plugin:terminal").projectDir = file("plugin/terminal")
+include(":plugin:testFixtures")
+project(":plugin:testFixtures").projectDir = file("plugin/testFixtures")
 include("lang.tt2")
 project(":lang.tt2").projectDir = file("tt2")
 include("lang.tt2:core")
